+++ spec/document_spec.rb
@@ -2,10 +2,15 @@ RSpec.describe Springcm::Document do
   let(:client) { Springcm::Client.new(data_center, client_id, client_secret) }
   let(:folder) { client.root_folder }
   let(:documents) { folder.documents }
+  let(:document) { documents.items.first }
 
   context "document API" do
     before(:each) do
       client.connect!
+    end
+
+    it "can be deleted" do
+      expect(document.delete).to be_a(Springcm::Document)
     end
 
     it "retrieves documents" do