+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to springcm-sdk will be documented in this file.
 
 ## [Unreleased]
+### Added
+* Folder#delete
+* Document#delete
 
 ## [0.3.3] - 2019-11-12
 ### Changed